[FLINK-21347][coordination] Extract interface from ExecutionGraph
