+++ flink-runtime/src/main/java/org/apache/flink/runtime/scheduler/SchedulerBase.java
@@ -46,12 +46,12 @@ import org.apache.flink.runtime.concurrent.FutureUtils;
 import org.apache.flink.runtime.deployment.TaskDeploymentDescriptorFactory;
 import org.apache.flink.runtime.execution.ExecutionState;
 import org.apache.flink.runtime.executiongraph.ArchivedExecutionGraph;
+import org.apache.flink.runtime.executiongraph.DefaultExecutionGraphBuilder;
 import org.apache.flink.runtime.executiongraph.ErrorInfo;
 import org.apache.flink.runtime.executiongraph.Execution;
 import org.apache.flink.runtime.executiongraph.ExecutionAttemptID;
 import org.apache.flink.runtime.executiongraph.ExecutionDeploymentListener;
 import org.apache.flink.runtime.executiongraph.ExecutionGraph;
-import org.apache.flink.runtime.executiongraph.ExecutionGraphBuilder;
 import org.apache.flink.runtime.executiongraph.ExecutionJobVertex;
 import org.apache.flink.runtime.executiongraph.ExecutionStateUpdateListener;
 import org.apache.flink.runtime.executiongraph.ExecutionVertex;
@@ -337,7 +337,7 @@ public abstract class SchedulerBase implements SchedulerNG, CheckpointScheduling
                     }
                 };
 
-        return ExecutionGraphBuilder.buildGraph(
+        return DefaultExecutionGraphBuilder.buildGraph(
                 jobGraph,
                 jobMasterConfiguration,
                 futureExecutor,
