+++ flink-runtime/src/test/java/org/apache/flink/runtime/executiongraph/DefaultExecutionGraphDeploymentTest.java
@@ -86,8 +86,8 @@ import static org.junit.Assert.assertNotNull;
 import static org.junit.Assert.assertThat;
 import static org.junit.Assert.fail;
 
-/** Tests for {@link ExecutionGraph} deployment. */
-public class ExecutionGraphDeploymentTest extends TestLogger {
+/** Tests for {@link DefaultExecutionGraph} deployment. */
+public class DefaultExecutionGraphDeploymentTest extends TestLogger {
 
     /** BLOB server instance to use for the job graph. */
     protected BlobWriter blobWriter = VoidBlobWriter.getInstance();
@@ -104,7 +104,7 @@ public class ExecutionGraphDeploymentTest extends TestLogger {
      *
      * @param eg the execution graph that was created
      */
-    protected void checkJobOffloaded(ExecutionGraph eg) throws Exception {
+    protected void checkJobOffloaded(DefaultExecutionGraph eg) throws Exception {
         assertTrue(eg.getJobInformationOrBlobKey().isLeft());
     }
 
@@ -152,8 +152,8 @@ public class ExecutionGraphDeploymentTest extends TestLogger {
                     v2, DistributionPattern.ALL_TO_ALL, ResultPartitionType.PIPELINED);
 
             DirectScheduledExecutorService executor = new DirectScheduledExecutorService();
-            ExecutionGraph eg =
-                    TestingExecutionGraphBuilder.newBuilder()
+            DefaultExecutionGraph eg =
+                    TestingDefaultExecutionGraphBuilder.newBuilder()
                             .setJobGraph(new JobGraph(jobId, "Test Job"))
                             .setFutureExecutor(executor)
                             .setIoExecutor(executor)
@@ -478,7 +478,7 @@ public class ExecutionGraphDeploymentTest extends TestLogger {
 
         final ExecutionGraph eg = scheduler.getExecutionGraph();
 
-        checkJobOffloaded(eg);
+        checkJobOffloaded((DefaultExecutionGraph) eg);
 
         // schedule, this triggers mock deployment
         scheduler.startScheduling();
@@ -535,7 +535,7 @@ public class ExecutionGraphDeploymentTest extends TestLogger {
                         .build();
         final ExecutionGraph eg = scheduler.getExecutionGraph();
 
-        checkJobOffloaded(eg);
+        checkJobOffloaded((DefaultExecutionGraph) eg);
 
         // schedule, this triggers mock deployment
         scheduler.startScheduling();
@@ -675,7 +675,7 @@ public class ExecutionGraphDeploymentTest extends TestLogger {
                                 0),
                         null));
 
-        return TestingExecutionGraphBuilder.newBuilder()
+        return TestingDefaultExecutionGraphBuilder.newBuilder()
                 .setJobGraph(jobGraph)
                 .setJobMasterConfig(configuration)
                 .setBlobWriter(blobWriter)
